Fix "array of simple types" case.
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/projects/Siemens/Soap/kdsoap/trunk@32691 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ kdsoap/kdwsdl2cpp/libkode/code.cpp
@@ -180,7 +180,7 @@ void Code::addFormattedText( const QString &text )
 
   QStringList::ConstIterator it;
   for ( it = words.constBegin(); it != words.constEnd(); ++it ) {
-    if ( (int)(*it).length() + lineLength >= maxWidth ) {
+    if ( (*it).length() + lineLength >= maxWidth ) {
       line = line.trimmed();
       addLine( line );
       line.truncate( 0 );
